% Generated by roxygen2: do not edit by hand
\docType{class}
\name{bitcoind}
\alias{bitcoind}
\title{bitcoind class}
\format{An R6 class object.}
\usage{
bitcoind
}
\description{
bitcoind class
}
\details{
Initiate object to manage all RPC methods.
}
\seealso{
\link{bitcoind.rpc}
}
\keyword{datasets}

