% Generated by roxygen2: do not edit by hand
\name{bitcoind.rpc}
\alias{bitcoind.rpc}
\title{Bitcoin daemon curl call wrapper.}
\usage{
bitcoind.rpc(host = getOption("rpchost", "127.0.0.1"),
  user = getOption("rpcuser"), password = getOption("rpcpassword"),
  port = getOption("rpcport", "8332"), id = NA_integer_, method,
  params = list())
}
\arguments{
\item{host}{character.}

\item{user}{character.}

\item{password}{character.}

\item{port}{character.}

\item{id}{any.}

\item{method}{character.}

\item{params}{list.}
}
\value{
jsonlite decoded response from json-rpc bitcoind. If \code{option("use.data.table"=TRUE)} then it will recursively turn data.frames to data.tables.
}
\description{
Bitcoin daemon curl call wrapper.
}
\examples{
\dontrun{
bitcoind.rpc(host = "127.0.0.1",
             user = "rpcuser",
             password = "rpcpassword",
             port = "18332", # regtest
             method = "getinfo")
}
}

