% Generated by roxygen2: do not edit by hand
\name{getaccountaddress}
\alias{getaccountaddress}
\title{getaccountaddress}
\usage{
getaccountaddress(account)
}
\arguments{
\item{account}{character.}
}
\description{
getaccountaddress
}
\seealso{
\url{https://bitcoin.org/en/developer-reference#getaccountaddress}
}

