% Generated by roxygen2: do not edit by hand
\name{getreceivedbyaccount}
\alias{getreceivedbyaccount}
\title{getreceivedbyaccount}
\usage{
getreceivedbyaccount(account, minconf = 1L)
}
\arguments{
\item{account}{character.}

\item{minconf}{integer.}
}
\description{
getreceivedbyaccount
}
\seealso{
\url{https://bitcoin.org/en/developer-reference#getreceivedbyaccount}
}

