% Generated by roxygen2: do not edit by hand
\name{getreceivedbyaddress}
\alias{getreceivedbyaddress}
\title{getreceivedbyaddress}
\usage{
getreceivedbyaddress(bitcoinaddress, minconf = 1L)
}
\arguments{
\item{bitcoinaddress}{character.}

\item{minconf}{integer.}
}
\description{
getreceivedbyaddress
}
\seealso{
\url{https://bitcoin.org/en/developer-reference#getreceivedbyaddress}
}

