% Generated by roxygen2: do not edit by hand
\name{gettransaction}
\alias{gettransaction}
\title{gettransaction}
\usage{
gettransaction(txid, includeWatchonly = FALSE)
}
\arguments{
\item{txid}{character.}

\item{includeWatchonly}{logical.}
}
\description{
gettransaction
}
\seealso{
\url{https://bitcoin.org/en/developer-reference#gettransaction}
}

