% Generated by roxygen2: do not edit by hand
\name{listunspent}
\alias{listunspent}
\title{listunspent}
\usage{
listunspent(miconf = 1L, maxconf = 9999999L, addresses)
}
\arguments{
\item{miconf}{integer.}

\item{maxconf}{integer.}

\item{addresses}{character.}
}
\description{
listunspent
}
\seealso{
\url{https://bitcoin.org/en/developer-reference#listunspent}
}

