% Generated by roxygen2: do not edit by hand
\name{makepaymenturi}
\alias{makepaymenturi}
\title{Generate payment uri.}
\usage{
makepaymenturi(address, amount)
}
\arguments{
\item{address}{character. Target address.}

\item{amount}{numeric. Amount to be hardcoded in payment uri.}
}
\value{
Character, payment uri link \code{bitcoin:ADDRESS&amount=0.1}.
}
\description{
Generate payment uri.
}

