% Generated by roxygen2: do not edit by hand
\docType{package}
\name{rbitcoind}
\alias{rbitcoind}
\alias{rbitcoind-package}
\title{R API to bitcoin daemon}
\description{
You need to setup/connect to bitcoin node. Unit tests are run on \code{regtest} network which uses private blockchain and doesn't need to be synchronized.
}
\examples{
\dontrun{
options(rpchost = "127.0.0.1",
        rpcuser = "bitcoinduser",
        rpcpassword = "userpassbitcoind",
        rpcport = "18332") # regtest
getinfo()
}
}
\author{
Jan Gorecki
}
\seealso{
\link{bitcoind.rpc}
}

