% Generated by roxygen2: do not edit by hand
\name{use.data.table}
\alias{use.data.table}
\title{Auto recursive setDT on data.frames}
\usage{
use.data.table(x, use = getOption("use.data.table"))
}
\arguments{
\item{x}{object to recursively transform.}

\item{use}{logical, default will keep the data.frames.}
}
\description{
Auto recursive setDT on data.frames
}
\details{
Requires data.table package.
}

