% Generated by roxygen2: do not edit by hand
\name{validateaddress}
\alias{validateaddress}
\title{validateaddress}
\usage{
validateaddress(bitcoinaddress)
}
\arguments{
\item{bitcoinaddress}{character.}
}
\description{
validateaddress
}
\seealso{
\url{https://bitcoin.org/en/developer-reference#validateaddress}
}

