% Generated by roxygen2: do not edit by hand
\docType{package}
\name{rbitcoind}
\alias{rbitcoind}
\alias{rbitcoind-package}
\alias{rbtcd}
\title{R interface to bitcoin daemon}
\description{
You need to have to bitcoin node installed, blockchain is not necessary. Unit tests runs on \emph{regtest} network which uses private blockchain and doesn't need to be synchronized.
}
\examples{
\dontrun{
# examples assumes your daemon is already running on regtest

# low level calls to json-rpc
bitcoind.rpc(user = "username",
             password = "password",
             port = "18332",
             method = "getinfo")

# bitcoin daemon object
btcd = bitcoind$new(rpcuser = "username",
                    rpcpassword = "password",
                    regtest = TRUE)
btcd$getinfo()
print(btcd)
}
}
\author{
Jan Gorecki
}
\references{
\url{https://bitcoin.org/en/glossary/regression-test-mode}
}
\seealso{
\link{bitcoind}, \link{bitcoind.rpc}
}

