% Generated by roxygen2: do not edit by hand
\docType{package}
\name{rbitcoind}
\alias{btcd}
\alias{rbitcoind}
\alias{rbitcoind-package}
\alias{rbtcd}
\title{R interface to bitcoin daemon}
\description{
You need to have to bitcoin node installed, blockchain is not necessary. Unit tests runs on \emph{regtest} network which uses private blockchain and doesn't need to be synchronized.
}
\examples{
\dontrun{
# low level calls to json-rpc
bitcoind.rpc(host = "127.0.0.1",
             user = "rpcuser",
             password = "rpcpassword",
             port = "18332", # regtest
             method = "getinfo")

# bitcoin daemon object
btcd <- bitcoind$new(rpcuser = "rpcuser",
                     rpcpassword = "rpcpassword",
                     regtest = TRUE)
btcd$getinfo()
}
}
\author{
Jan Gorecki
}
\references{
\url{https://bitcoin.org/en/glossary/regression-test-mode}
}
\seealso{
\link{bitcoind}, \link{bitcoind.rpc}
}

